refactor: OrderCheckPage 도메인 로직 함수로 분리
